drv-iio-compass: Change the compass channel name
The _raw is not needed in the channel name for process_scan_1(). This function looks up the channel by it's channel name or in this case "in_rot_from_north_magnetic_tilt_comp".
The channel name is read from the file ending in _en from the scanning_element folder, in build_channel_array().